Cover fork edge cases after workspace switches
