+++ Sources/RestorableAgentSession.swift
@@ -1,6 +1,5 @@
 import Foundation
 import CMUXAgentLaunch
-import Darwin
 
 fileprivate func shellSingleQuoted(_ value: String) -> String {
     "'" + value.replacingOccurrences(of: "'", with: "'\\''") + "'"
@@ -483,260 +482,6 @@ enum AgentResumeCommandBuilder {
             return nil
         }
         return trimmed
-    }
-}
-
-private actor OpenCodeVersionProbeCache {
-    private var valuesByKey: [String: Bool] = [:]
-
-    func value(for key: String) -> Bool? {
-        valuesByKey[key]
-    }
-
-    func store(_ value: Bool, for key: String) {
-        valuesByKey[key] = value
-    }
-}
-
-enum AgentForkSupport {
-    static let minimumOpenCodeForkVersion = SemanticVersion(major: 1, minor: 14, patch: 50)
-    private static let commandOutputTimeoutNanoseconds: Int64 = 3_000_000_000
-    private static let commandTerminateTimeoutNanoseconds: Int64 = 500_000_000
-    private static let openCodeVersionProbeCache = OpenCodeVersionProbeCache()
-
-    private final class CommandOutputBuffer: @unchecked Sendable {
-        private let lock = NSLock()
-        private var data = Data()
-
-        func append(_ chunk: Data) {
-            lock.lock()
-            data.append(chunk)
-            lock.unlock()
-        }
-
-        func value() -> Data {
-            lock.lock()
-            let snapshot = data
-            lock.unlock()
-            return snapshot
-        }
-    }
-
-    static func supportsFork(snapshot: SessionRestorableAgentSnapshot) async -> Bool {
-        guard snapshot.forkCommand != nil else { return false }
-        guard snapshot.kind == .opencode else { return true }
-        if snapshot.launchCommand?.launcher == "omo" {
-            return true
-        }
-        guard let probe = AgentResumeCommandBuilder.openCodeVersionProbe(
-            launchCommand: snapshot.launchCommand
-        ) else {
-            return false
-        }
-        let workingDirectory = openCodeProbeWorkingDirectory(snapshot: snapshot)
-        guard shouldRunLocalOpenCodeVersionProbe(
-            probe: probe,
-            workingDirectory: workingDirectory
-        ) else {
-            return true
-        }
-        let cacheKey = openCodeVersionProbeCacheKey(
-            probe: probe,
-            environment: snapshot.launchCommand?.environment,
-            workingDirectory: workingDirectory
-        )
-        if let cached = await openCodeVersionProbeCache.value(for: cacheKey) {
-            return cached
-        }
-        guard let output = await commandOutput(
-            executable: probe.executable,
-            arguments: probe.arguments,
-            environment: snapshot.launchCommand?.environment,
-            workingDirectory: workingDirectory
-        ) else {
-            await openCodeVersionProbeCache.store(false, for: cacheKey)
-            return false
-        }
-        let supportsFork = openCodeVersionSupportsFork(output)
-        await openCodeVersionProbeCache.store(supportsFork, for: cacheKey)
-        return supportsFork
-    }
-
-    static func openCodeVersionSupportsFork(_ output: String) -> Bool {
-        guard let version = SemanticVersion.first(in: output) else {
-            return false
-        }
-        return version >= minimumOpenCodeForkVersion
-    }
-
-    private static func openCodeVersionProbeCacheKey(
-        probe: (executable: String, arguments: [String]),
-        environment: [String: String]?,
-        workingDirectory: String?
-    ) -> String {
-        var processEnvironment = ProcessInfo.processInfo.environment
-        if let environment {
-            let selectedEnvironment = AgentLaunchEnvironmentPolicy.selectedEnvironment(from: environment)
-            for (key, value) in selectedEnvironment {
-                processEnvironment[key] = value
-            }
-        }
-        let relevantEnvironmentKeys = [
-            "PATH",
-            "OPENCODE_BIN",
-            "OPENCODE_CONFIG_DIR"
-        ]
-        let environmentParts = relevantEnvironmentKeys.map { key in
-            "\(key)=\(processEnvironment[key] ?? "")"
-        }
-        return ([probe.executable] + probe.arguments + environmentParts + ["cwd=\(workingDirectory ?? "")"])
-            .joined(separator: "\u{1f}")
-    }
-
-    private static func commandOutput(
-        executable: String,
-        arguments: [String],
-        environment: [String: String]?,
-        workingDirectory: String?
-    ) async -> String? {
-        await Task.detached(priority: .utility) {
-            commandOutputSynchronously(
-                executable: executable,
-                arguments: arguments,
-                environment: environment,
-                workingDirectory: workingDirectory
-            )
-        }.value
-    }
-
-    private static func commandOutputSynchronously(
-        executable: String,
-        arguments: [String],
-        environment: [String: String]?,
-        workingDirectory: String?
-    ) -> String? {
-        let process = Process()
-        process.executableURL = URL(fileURLWithPath: "/usr/bin/env")
-        process.arguments = [executable] + arguments
-        if let workingDirectoryURL = localDirectoryURL(path: workingDirectory) {
-            process.currentDirectoryURL = workingDirectoryURL
-        }
-
-        let pipe = Pipe()
-        process.standardOutput = pipe
-        process.standardError = pipe
-        let outputBuffer = CommandOutputBuffer()
-        pipe.fileHandleForReading.readabilityHandler = { handle in
-            let data = handle.availableData
-            guard !data.isEmpty else { return }
-            outputBuffer.append(data)
-        }
-
-        var processEnvironment = ProcessInfo.processInfo.environment
-        if let environment {
-            let selectedEnvironment = AgentLaunchEnvironmentPolicy.selectedEnvironment(from: environment)
-            for (key, value) in selectedEnvironment {
-                processEnvironment[key] = value
-            }
-        }
-        process.environment = processEnvironment
-        let completion = DispatchSemaphore(value: 0)
-        process.terminationHandler = { _ in completion.signal() }
-
-        do {
-            try process.run()
-        } catch {
-            pipe.fileHandleForReading.readabilityHandler = nil
-            return nil
-        }
-        var timedOut = false
-        if completion.wait(timeout: .now() + .nanoseconds(Int(commandOutputTimeoutNanoseconds))) == .timedOut {
-            timedOut = true
-            process.terminate()
-            if completion.wait(timeout: .now() + .nanoseconds(Int(commandTerminateTimeoutNanoseconds))) == .timedOut {
-                kill(process.processIdentifier, SIGKILL)
-                completion.wait()
-            }
-        }
-
-        pipe.fileHandleForReading.readabilityHandler = nil
-        let remainingData = pipe.fileHandleForReading.readDataToEndOfFile()
-        outputBuffer.append(remainingData)
-        guard !timedOut else { return nil }
-        return String(data: outputBuffer.value(), encoding: .utf8)
-    }
-
-    private static func openCodeProbeWorkingDirectory(snapshot: SessionRestorableAgentSnapshot) -> String? {
-        normalized(snapshot.launchCommand?.workingDirectory) ?? normalized(snapshot.workingDirectory)
-    }
-
-    private static func shouldRunLocalOpenCodeVersionProbe(
-        probe: (executable: String, arguments: [String]),
-        workingDirectory: String?
-    ) -> Bool {
-        if let workingDirectory, localDirectoryURL(path: workingDirectory) == nil {
-            return false
-        }
-        if probe.executable.hasPrefix("/") {
-            return FileManager.default.isExecutableFile(atPath: probe.executable)
-        }
-        return true
-    }
-
-    private static func localDirectoryURL(path: String?) -> URL? {
-        guard let path = normalized(path) else { return nil }
-        var isDirectory: ObjCBool = false
-        guard FileManager.default.fileExists(atPath: path, isDirectory: &isDirectory),
-              isDirectory.boolValue else {
-            return nil
-        }
-        return URL(fileURLWithPath: path, isDirectory: true)
-    }
-
-    private static func normalized(_ value: String?) -> String? {
-        guard let trimmed = value?.trimmingCharacters(in: .whitespacesAndNewlines),
-              !trimmed.isEmpty else {
-            return nil
-        }
-        return trimmed
-    }
-}
-
-struct SemanticVersion: Comparable, Sendable {
-    let major: Int
-    let minor: Int
-    let patch: Int
-
-    static func < (lhs: SemanticVersion, rhs: SemanticVersion) -> Bool {
-        if lhs.major != rhs.major { return lhs.major < rhs.major }
-        if lhs.minor != rhs.minor { return lhs.minor < rhs.minor }
-        return lhs.patch < rhs.patch
-    }
-
-    static func first(in output: String) -> SemanticVersion? {
-        let pattern = #"(\d+)\.(\d+)(?:\.(\d+))?"#
-        guard let expression = try? NSRegularExpression(pattern: pattern) else {
-            return nil
-        }
-        let range = NSRange(output.startIndex..<output.endIndex, in: output)
-        guard let match = expression.firstMatch(in: output, range: range) else {
-            return nil
-        }
-
-        func integer(at captureIndex: Int, fallback defaultValue: Int? = nil) -> Int? {
-            let captureRange = match.range(at: captureIndex)
-            guard captureRange.location != NSNotFound,
-                  let range = Range(captureRange, in: output) else {
-                return defaultValue
-            }
-            return Int(output[range])
-        }
-
-        guard let major = integer(at: 1),
-              let minor = integer(at: 2) else {
-            return nil
-        }
-        return SemanticVersion(major: major, minor: minor, patch: integer(at: 3, fallback: 0) ?? 0)
     }
 }
